Basic CRUD, rendering, aliases system
Now accessible two variants of rendering templates: execute & render.
First is execute a php file, second is read a file and render it (change
[[name]] elements to data). Also we provide you the aliases system
(short names of some function - render may be use as 'r', execute - as
'e').
Finally, we plug a simple, but flexible CRUD system. It functionality
will be extended in future.
Fresh mysql dump - in a folder mysql_dump.